Working update

+++ frc-2019-robot/src/main/cpp/routine/set_elevator_position_routine.cpp
@@ -3,10 +3,10 @@
 #include <robot.hpp>
 
 namespace garage {
-    SetElevatorPositionRoutine::SetElevatorPositionRoutine(std::shared_ptr<Robot> robot, int setPoint, const std::string& name)
+    SetElevatorPositionRoutine::SetElevatorPositionRoutine(std::shared_ptr<Robot> robot, double setPoint, const std::string& name)
             : SubsystemRoutine(robot, robot->GetElevator(), name), m_SetPoint(setPoint) {
         if (m_Subsystem) {
-            m_Subsystem->Log(lib::Logger::LogLevel::k_Verbose, lib::Logger::Format("[%s] Set Elevator Set Point: %d", FMT_STR(name), setPoint));
+            m_Subsystem->Log(lib::Logger::LogLevel::k_Verbose, lib::Logger::Format("[%s] Set Elevator Set Point: %f", FMT_STR(name), setPoint));
         }
     }
