+++ frc-2019-robot/src/main/include/routine/elevator_and_flipper_routine.hpp
@@ -8,7 +8,7 @@
 namespace garage {
     class ElevatorAndFlipperRoutine : public lib::ParallelRoutine {
     public:
-        ElevatorAndFlipperRoutine(std::shared_ptr<Robot> robot, int setPoint, double angle, const std::string& name = "Elevator And Flipper Routine")
+        ElevatorAndFlipperRoutine(std::shared_ptr<Robot> robot, double setPoint, double angle, const std::string& name = "Elevator And Flipper Routine")
                 : lib::ParallelRoutine(robot, name, {
                 std::make_shared<SetElevatorPositionRoutine>(robot, setPoint),
                 std::make_shared<SetFlipperAngleRoutine>(robot, angle)
